requirements.txt
numpy>=1.26,<2
torch>=2.0
xarray
netCDF4
matplotlib
PyYAML
scipy
